refactor: moved user popover to header folder gf-58
